pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uai"
version = "0.1.0"
description = "UAI - Unified AI CLI: one tool for all AI providers with persistent context, intelligent routing and automatic fallback"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Diego Camara" }]
keywords = ["ai", "cli", "claude", "gemini", "codex", "qwen", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "typer>=0.12.0",
    "prompt_toolkit>=3.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "keyring>=25.0",
    "cryptography>=43.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0",
    "openai>=1.60.0",
    "httpx>=0.27.0",
    "structlog>=24.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
privacy = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
uai = "uai.cli.app:main"

[project.entry-points."uai.providers"]
# Third-party providers register here:
# my-provider = "my_package:MyProvider"

[tool.hatch.build.targets.wheel]
packages = ["src/uai"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
